requirements.txt
aiohttp==3.9.5
aiohttp-socks==0.8.4
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
iniconfig==2.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pluggy==1.5.0
Pygments==2.18.0
pytest==8.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-socks==2.4.4
pytz==2024.1
requests==2.32.2
rich==13.7.1
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.30
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
